smooth resizing v2
improvements to previous "beta" commit. Resizing the window manually is now smooth for the most part.

lines removed from the wiki:
**Resizing smoothness**
Snapping the window to Windows 11 snap layouts is the smoothest method.
Manual resizing of the window can be stuttery, efforts are being made to make this smoother by adding delays to certain panels' reactions.